Minor refactor of code into seperate files. Needs moar
